Const auf 64 bit erweitert

+++ knf_gen/common/solvertools.h
@@ -5,12 +5,12 @@
 
 #include "cryptominisat4/cryptominisat.h"
 
-uint32_t solver_readInt(CMSat::SATSolver& solver, unsigned lsb, unsigned bitWidth);
+uint64_t solver_readInt(CMSat::SATSolver& solver, unsigned lsb, unsigned bitWidth);
 
-void solver_writeInt(CMSat::SATSolver& solver, unsigned lsb, unsigned bitWidth, unsigned value);
+void solver_writeInt(CMSat::SATSolver& solver, unsigned lsb, unsigned bitWidth, uint64_t value);
 
-uint32_t solver_readInt_msb(CMSat::SATSolver& solver, unsigned msb, unsigned bitWidth);
+uint64_t solver_readInt_msb(CMSat::SATSolver& solver, unsigned msb, unsigned bitWidth);
 
-void solver_writeInt_msb(CMSat::SATSolver& solver, unsigned msb, unsigned bitWidth, unsigned value);
+void solver_writeInt_msb(CMSat::SATSolver& solver, unsigned msb, unsigned bitWidth, uint64_t value);
 
 #endif //__SOLVERTOOLS_H__
